backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
openpyxl==3.1.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-docx==0.8.11
numpy==1.26.0
pydantic==2.4.2
